Functional unit tests and from the UI
got the system initialization and config for the store and the payment gateway working.

and returning a properly structured json message to the ui.

Both the unit tests pass and a create payment method from the ui returns valid json.

Now have to map the input data and start mapping to mimi and userstore.