Using Real Data all the way to paygate
The data passed in the request, If valid, passes all the way to paygate and back.

The error handling is not working though if the data is not valid.

I am throwing an exception from the payment_gateway (throw (ex-info ...)) but the store still renders a response. I don't quite understand this.